requirements.txt
tensorflow
fastapi
numpy
os
logging
pillow